Fix TS errors with filtered evm configs in the components. Move external config initialization in the main App component. Move an active actions function inside actions index.ts
